Fixed Bugs 5628, 5691, 5693, 5544 Made sure the rules were processed after creating a new object (they weren't being called) Fixed the LatLonUI Plugin to better notify of changes, incomplete data and errors. Additional update to WebLinks UIValidators can now provide text as to what the error was.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@4400 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/ui/forms/FormViewObj.java
@@ -2586,24 +2586,30 @@ public class FormViewObj implements Viewable,
                 Integer          objId = fdo.getId();
                 if (objId != null)
                 {
-                    session.beginTransaction();
-                    
                     // Reload the object from the database  to avoid a stale object exception.
                     Object dbDataObj = session.getData(fdo.getDataClass(), "id", objId, DataProviderSessionIFace.CompareType.Equals);
-                    session.delete(dbDataObj);
-                    if (businessRules != null)
+                    if (dbDataObj != null)
                     {
-                        if (!businessRules.beforeDeleteCommit(dbDataObj, session))
+                        session.beginTransaction();
+                        session.delete(dbDataObj);
+                        if (businessRules != null)
                         {
-                            session.rollback();
-                            throw new Exception("Business rules processing failed");
+                            if (!businessRules.beforeDeleteCommit(dbDataObj, session))
+                            {
+                                session.rollback();
+                                throw new Exception("Business rules processing failed");
+                            }
                         }
-                    }
-                    session.commit();
-                    session.flush();
-                    if (businessRules != null)
+                        session.commit();
+                        session.flush();
+                        if (businessRules != null)
+                        {
+                            businessRules.afterDeleteCommit(dbDataObj);
+                        }
+                    } else
                     {
-                        businessRules.afterDeleteCommit(dbDataObj);
+                        doClearObj = true;
+                        UIRegistry.showLocalizedMsg("Warning", "OBJ_ALREADY_DEL");
                     }
                     
                     isNewlyCreatedDataObj = false; // shouldn't be needed, but just in case
@@ -2615,6 +2621,7 @@ public class FormViewObj implements Viewable,
                     {
                         formValidator.setNewObj(isNewlyCreatedDataObj);
                     }
+
                 }
                 
             } catch (edu.ku.brc.dbsupport.StaleObjectException e)
@@ -2791,6 +2798,10 @@ public class FormViewObj implements Viewable,
                 {
                     UIValidator.setIgnoreAllValidation(this, true);
                     createNewDataObject(false);
+                    if (formValidator != null)
+                    {
+                        formValidator.processFormRules();
+                    }
                     UIValidator.setIgnoreAllValidation(this, false);
                     //focusFirstFormControl();
                 }
